requirements.txt
fastapi
uvicorn
pyvespa
python-multipart
langchain
langchain-community
pydantic
openpyxl 
python-docx 
pypdf
unstructured
sentence-transformers>=2.6.0
transformers>=4.40.0
huggingface_hub>=0.20.0